moved ViewInfoModal.tsx to shared
