+++ ui/components/MesheryAdapterPlayComponent.tsx
@@ -47,7 +47,7 @@ import fetchAvailableNamespaces from '@/graphql/queries/NamespaceQuery';
 import MesheryMetrics from './Performance/MesheryMetrics';
 import MesheryResultDialog from './MesheryResultDialog';
 import ReactSelectWrapper from './ReactSelectWrapper';
-import ConfirmationMsg from './ConfirmationModal';
+import ConfirmationMsg from './shared/Modal/ConfirmationModal';
 import { iconMedium } from '../css/icons.styles';
 import { ACTIONS } from '../utils/Enum';
 import { getModelByName } from '../api/meshmodel';
